🚚 Renamed `Header` layout component
